@@ -1,5 +1,5 @@
 # This migration comes from acts_as_taggable_on_engine (originally 6)
-class AddMissingIndexes < ActiveRecord::Migration
+class AddMissingIndexes < ActiveRecord::Migration[5.0]
   def change
     add_index :taggings, :tag_id
     add_index :taggings, :taggable_id
